0.6 - Alpha 4
Added fire and water to the sails and fuel. Still not sure where to fit it in for the motor. Also need to figure out how to repair. Make a whole new hose repair thing? Who knows.
Progress made on replicating the SM511/512 audio chip, with a preset made in ableton emulating the melody chip from those processors. Will need to take it further to actually create sounds.

Still to complete:
- Fire & Water for motor
- Moving Backgrounds
- Obstacles & Damage
- Repairs
- End State?
- Audio