0.86 - Input Generalization
Generalized input for keyboard or gameplay. It's not pretty. But I switched to Unity's new Input System and it was a huge improvement in terms of managing the inputs, so I'll call it a win. Should be complete, but there may be a bug hiding somewhere. Also gamepad play is almost impossible.

Still to complete:
- Welding sprites
- Moving Backgrounds
- High Obstacles
- End State
- Audio